Reorganise factories to match the models
Much like our models layout this sort of ignores applications, but
otherwise moves each factory into its own module, maintaing the
namespacing already present in the __init__.py.
