+++ tests/factories/__init__.py
@@ -1,6 +1,24 @@
-from .applications import *  # noqa: F401, F403
-from .interactive import *  # noqa: F401, F403
-from .jobserver import *  # noqa: F401, F403
-from .redirects import *  # noqa: F401, F403
-from .releases import *  # noqa: F401, F403
-from .reports import *  # noqa: F401, F403
+from .analysis_request import *  # noqa: F401, F403
+from .application import *  # noqa: F401, F403
+from .auditable_event import *  # noqa: F401, F403
+from .backend import *  # noqa: F401, F403
+from .backend_membership import *  # noqa: F401, F403
+from .job import *  # noqa: F401, F403
+from .job_request import *  # noqa: F401, F403
+from .org import *  # noqa: F401, F403
+from .org_membership import *  # noqa: F401, F403
+from .partial import *  # noqa: F401, F403
+from .project import *  # noqa: F401, F403
+from .project_collaboration import *  # noqa: F401, F403
+from .publish_request import *  # noqa: F401, F403
+from .redirect import *  # noqa: F401, F403
+from .release import *  # noqa: F401, F403
+from .release_file import *  # noqa: F401, F403
+from .release_file_review import *  # noqa: F401, F403
+from .repo import *  # noqa: F401, F403
+from .report import *  # noqa: F401, F403
+from .snapshot import *  # noqa: F401, F403
+from .stats import *  # noqa: F401, F403
+from .user import *  # noqa: F401, F403
+from .user_social_auth import *  # noqa: F401, F403
+from .workspace import *  # noqa: F401, F403
