+++ tests/factories/auditable_event.py
@@ -1,0 +1,8 @@
+import factory
+
+from jobserver.models import AuditableEvent
+
+
+class AuditableEventFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = AuditableEvent
