+++ tests/factories/backend_membership.py
@@ -1,0 +1,13 @@
+import factory
+
+from jobserver.models import BackendMembership
+
+
+class BackendMembershipFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = BackendMembership
+
+    backend = factory.SubFactory("tests.factories.BackendFactory")
+    user = factory.SubFactory("tests.factories.UserFactory")
+
+    created_by = factory.SubFactory("tests.factories.UserFactory")
