+++ tests/factories/job_request.py
@@ -1,0 +1,14 @@
+import factory
+
+from jobserver.models import JobRequest
+
+
+class JobRequestFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = JobRequest
+
+    backend = factory.SubFactory("tests.factories.BackendFactory")
+    created_by = factory.SubFactory("tests.factories.UserFactory")
+    workspace = factory.SubFactory("tests.factories.WorkspaceFactory")
+
+    requested_actions = []
