+++ tests/factories/org_membership.py
@@ -1,0 +1,11 @@
+import factory
+
+from jobserver.models import OrgMembership
+
+
+class OrgMembershipFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = OrgMembership
+
+    org = factory.SubFactory("tests.factories.OrgFactory")
+    user = factory.SubFactory("tests.factories.UserFactory")
