+++ tests/factories/stats.py
@@ -1,0 +1,14 @@
+from datetime import UTC
+
+import factory
+
+from jobserver.models import Stats
+
+
+class StatsFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = Stats
+
+    backend = factory.SubFactory("tests.factories.BackendFactory")
+
+    api_last_seen = factory.Faker("date_time", tzinfo=UTC)
